module6-task2 | fix homework: список сортировки вынесен в отдельный компонент
